First tests implemented for config_flow
