+++ homeassistant/components/homematicip_cloud/__init__.py
@@ -16,7 +16,8 @@ from .const import (
     CONF_ACCESSPOINT, CONF_AUTHTOKEN, CONF_NAME)
 # Loading the config flow file will register the flow
 from .config_flow import configured_haps
-from .hap import HomematicipHAP, HomematicipGenericDevice  # noqa: F401
+from .hap import HomematicipHAP
+from .generic import HomematicipGenericDevice  # noqa: F401
 
 REQUIREMENTS = ['homematicip==0.9.6']
