+++ homeassistant/components/homematicip_cloud/hap.py
@@ -10,30 +10,16 @@ import logging
 
 from homeassistant import config_entries
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.entity import Entity
 from homeassistant.core import callback
+
+import homematicip
+from homematicip.base.base_connection import HmipConnectionError
 
 from .const import (
     HMIPC_HAPID, HMIPC_AUTHTOKEN, HMIPC_NAME,
     COMPONENTS)
 
 _LOGGER = logging.getLogger(__name__)
-
-ATTR_HOME_ID = 'home_id'
-ATTR_HOME_NAME = 'home_name'
-ATTR_DEVICE_ID = 'device_id'
-ATTR_DEVICE_LABEL = 'device_label'
-ATTR_STATUS_UPDATE = 'status_update'
-ATTR_FIRMWARE_STATE = 'firmware_state'
-ATTR_UNREACHABLE = 'unreachable'
-ATTR_LOW_BATTERY = 'low_battery'
-ATTR_MODEL_TYPE = 'model_type'
-ATTR_GROUP_TYPE = 'group_type'
-ATTR_DEVICE_RSSI = 'device_rssi'
-ATTR_DUTY_CYCLE = 'duty_cycle'
-ATTR_CONNECTED = 'connected'
-ATTR_SABOTAGE = 'sabotage'
-ATTR_OPERATION_LOCK = 'operation_lock'
 
 
 class HomematicipHAP(object):
@@ -54,7 +40,6 @@ class HomematicipHAP(object):
     async def async_setup(self, tries=0):
         """Initialize connection."""
         from homematicip.base.base_connection import HmipConnectionError
-
         try:
             self.home = await self.get_hap(
                 self.hass,
@@ -118,8 +103,6 @@ class HomematicipHAP(object):
 
     def get_state_finished(self, future):
         """Execute when get_state coroutine has finished."""
-        from homematicip.base.base_connection import HmipConnectionError
-
         try:
             future.result()
         except HmipConnectionError:
@@ -142,7 +125,6 @@ class HomematicipHAP(object):
 
     async def _handle_connection(self):
         """Handle websocket connection."""
-        from homematicip.base.base_connection import HmipConnectionError
         try:
             await self.home.get_current_state()
         except HmipConnectionError:
@@ -155,8 +137,6 @@ class HomematicipHAP(object):
 
     async def async_connect(self):
         """Start websocket connection."""
-        from homematicip.base.base_connection import HmipConnectionError
-
         tries = 0
         while True:
             try:
@@ -215,52 +195,3 @@ class HomematicipHAP(object):
         hass.loop.create_task(self.async_connect())
 
         return home
-
-
-class HomematicipGenericDevice(Entity):
-    """Representation of an HomematicIP generic device."""
-
-    def __init__(self, home, device, post=None):
-        """Initialize the generic device."""
-        self._home = home
-        self._device = device
-        self.post = post
-        _LOGGER.info('Setting up %s (%s)', self.name,
-                     self._device.modelType)
-
-    async def async_added_to_hass(self):
-        """Register callbacks."""
-        self._device.on_update(self._device_changed)
-
-    def _device_changed(self, json, **kwargs):
-        """Handle device state changes."""
-        _LOGGER.debug('Event %s (%s)', self.name, self._device.modelType)
-        self.async_schedule_update_ha_state()
-
-    @property
-    def name(self):
-        """Return the name of the generic device."""
-        name = self._device.label
-        if (self._home.name is not None and self._home.name != ''):
-            name = "{} {}".format(self._home.name, name)
-        if (self.post is not None and self.post != ''):
-            name = "{} {}".format(name, self.post)
-        return name
-
-    @property
-    def should_poll(self):
-        """No polling needed."""
-        return False
-
-    @property
-    def available(self):
-        """Device available."""
-        return not self._device.unreach
-
-    @property
-    def device_state_attributes(self):
-        """Return the state attributes of the generic device."""
-        return {
-            ATTR_LOW_BATTERY: self._device.lowBat,
-            ATTR_MODEL_TYPE: self._device.modelType
-        }
